pfr/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cymem==2.0.8
dataclasses-json==0.6.4
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.109.2
frozenlist==1.4.1
ghp-import==2.1.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
isodate==0.6.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langcodes==3.3.0
langsmith==0.1.23
Markdown==3.5.2
MarkupSafe==2.1.5
marshmallow==3.21.1
mergedeep==1.3.4
mkdocs==1.5.3
mkdocs-material==9.5.9
mkdocs-material-extensions==1.3.1
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
neo4j==5.18.0
numpy==1.26.4
openai==1.13.3
orjson==3.9.15
packaging==23.2
paginate==0.5.6
pathspec==0.12.1
platformdirs==4.2.0
preshed==3.0.9
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
pymdown-extensions==10.7
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyyaml_env_tag==0.1
rdflib==7.0.0
redis==5.0.2
regex==2023.12.25
requests==2.31.0
setuptools==69.2.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.28
srsly==2.4.8
starlette==0.36.3
tenacity==8.2.3
thinc==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.10.0
Unidecode==1.3.8
urllib3==2.2.1
uvicorn==0.27.1
wasabi==1.1.2
watchdog==4.0.0
weasel==0.3.4
yarl==1.9.4
